Refactor : utils와 api 폴더 구조 분리
